backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Protein Data Analysis Backend with MongoDB and Neo4j"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pymongo>=4.6.0",
    "motor>=3.3.0",
    "neo4j>=5.14.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "networkx>=3.2.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
